implemented endpoint that uses data demo
also refactored code into separate files
