+++ backend/src/db_helper.py
@@ -1,0 +1,17 @@
+import sqlite3
+
+from flask import g
+
+
+DATABASE = "database.db"
+
+def get_db():
+    db = getattr(g, '_database', None)
+    if db is None:
+        db = g._database = sqlite3.connect(DATABASE)
+    return db
+
+def close_db():
+	db = getattr(g, '_database', None)
+	if db is not None:
+		db.close()
