+++ frontend/src/lib/api.ts
@@ -1,6 +1,6 @@
 import fetch from 'cross-fetch'
 
-const API_URL = 'http://localhost:5050/'
+const API_URL = 'http://localhost:5050/db-demo'
 
 const getQuery = async () => {
   const response = await fetch(`${API_URL}`)
